requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-google-genai
langchain
langchain-core